planner-harness/requirements.txt
openai
PyYAML
rich